EGS-983: Switch to regular int for numbers
"32 bits should be enough for anybody!"

+++ engine/src/rhs_types/int.rs
@@ -2,15 +2,15 @@ use lex::{expect, span, take_while, Lex, LexErrorKind, LexResult};
 use std::ops::RangeInclusive;
 use strict_partial_ord::StrictPartialOrd;
 
-fn number(input: &str, radix: u32) -> LexResult<u64> {
+fn number(input: &str, radix: u32) -> LexResult<i32> {
     let (digits, input) = take_while(input, "digit", |c| c.is_digit(radix))?;
-    match u64::from_str_radix(digits, radix) {
+    match i32::from_str_radix(digits, radix) {
         Ok(res) => Ok((res, input)),
         Err(err) => Err((LexErrorKind::ParseInt { err, radix }, digits)),
     }
 }
 
-impl<'i> Lex<'i> for u64 {
+impl<'i> Lex<'i> for i32 {
     fn lex(input: &str) -> LexResult<Self> {
         if let Ok(input) = expect(input, "0") {
             match input.chars().next() {
@@ -24,12 +24,12 @@ impl<'i> Lex<'i> for u64 {
     }
 }
 
-impl<'i> Lex<'i> for RangeInclusive<u64> {
+impl<'i> Lex<'i> for RangeInclusive<i32> {
     fn lex(input: &str) -> LexResult<Self> {
         let initial_input = input;
-        let (first, input) = u64::lex(input)?;
+        let (first, input) = i32::lex(input)?;
         let (last, input) = if let Ok(input) = expect(input, "..") {
-            u64::lex(input)?
+            i32::lex(input)?
         } else {
             (first, input)
         };
@@ -43,21 +43,21 @@ impl<'i> Lex<'i> for RangeInclusive<u64> {
     }
 }
 
-impl StrictPartialOrd for u64 {}
+impl StrictPartialOrd for i32 {}
 
 #[test]
 fn test() {
-    assert_ok!(u64::lex("0"), 0u64, "");
-    assert_ok!(u64::lex("0-"), 0u64, "-");
-    assert_ok!(u64::lex("0x1f5+"), 501u64, "+");
-    assert_ok!(u64::lex("0123;"), 83u64, ";");
-    assert_ok!(u64::lex("78!"), 78u64, "!");
-    assert_ok!(u64::lex("0xefg"), 239u64, "g");
-    assert_ok!(RangeInclusive::lex("78!"), 78u64..=78u64, "!");
-    assert_ok!(RangeInclusive::lex("0..10"), 0u64..=10u64);
-    assert_ok!(RangeInclusive::lex("0123..0xefg"), 83u64..=239u64, "g");
+    assert_ok!(i32::lex("0"), 0i32, "");
+    assert_ok!(i32::lex("0-"), 0i32, "-");
+    assert_ok!(i32::lex("0x1f5+"), 501i32, "+");
+    assert_ok!(i32::lex("0123;"), 83i32, ";");
+    assert_ok!(i32::lex("78!"), 78i32, "!");
+    assert_ok!(i32::lex("0xefg"), 239i32, "g");
+    assert_ok!(RangeInclusive::lex("78!"), 78i32..=78i32, "!");
+    assert_ok!(RangeInclusive::lex("0..10"), 0i32..=10i32);
+    assert_ok!(RangeInclusive::lex("0123..0xefg"), 83i32..=239i32, "g");
     assert_err!(
-        <RangeInclusive<u64>>::lex("10..0"),
+        <RangeInclusive<i32>>::lex("10..0"),
         LexErrorKind::IncompatibleRangeBounds,
         "10..0"
     );
